\name{RcppTestPackage-package}
\alias{RcppTestPackage-package}
\alias{RcppTestPackage}
\docType{package}
\title{\packageTitle{RcppTestPackage}}
\description{\packageDescription{RcppTestPackage}}
\section{Package Content}{\packageIndices{RcppTestPackage}}
\author{\packageAuthor{RcppTestPackage}}
\section{Maintainer}{\packageMaintainer{RcppTestPackage}}
\keyword{package}
